refactor bert truncated as subclass
